Added helpers for methods' argument checking. Fixed a issue in the testing tool. Removed the clvs file.
The helpers can be used by including the types/typeutils.h file. Its
recommendable to use the macro CLEVER_CHECK_ARGS("Class::methodName",
CLEVER_TYPE("Type1"), CLEVER_TYPE("Type2"), …, CLEVER_TYPE("TypeN"),
NULL) which receives the method name string and the argument types **in
order** and a NULL constant to mark the end. The compiler shows a
warning because of variadic macros was included in C99 (because of gcc
-ansi flag)