Fixed Glob support for Windows (MSVC)
The equivalent to glob_t in Windows seems to be FindFirstFile and
others, so I implemented it instead of creating a CGlob class. The
reason why I did this way is because it's simple and the results are
loaded in the Glob class constructor, and are not controlled by the
user, which means there's no need to keep the handler everywhere.

I didn't test on other platforms yet, though. As soon as someone tests
this on a *nix environment I'll merge this to the upstream.