- Removed uneeded "outer global scope"
Before this change the symbol resolver would create "two global scopes",
one for the built in types and functions defined in std.core and other
for the program. The first was explicitly instantiated and the later
created by the "program block".

This change improves the situation by using the "program block" escope
as global escope.